Rename `handler` -> `handlers`
