+++ aiogram/dispatcher/event/handler.py
@@ -8,7 +8,7 @@ from typing import Any, Callable, Dict, List, Optional, Tuple
 from magic_filter import MagicFilter
 
 from aiogram.dispatcher.flags import extract_flags_from_object
-from aiogram.handler import BaseHandler
+from aiogram.handlers import BaseHandler
 
 CallbackType = Callable[..., Any]
