+++ tests/test_dispatcher/test_event/test_handler.py
@@ -6,7 +6,7 @@ import pytest
 from aiogram import F
 from aiogram.dispatcher.event.handler import CallableMixin, FilterObject, HandlerObject
 from aiogram.filters import BaseFilter
-from aiogram.handler import BaseHandler
+from aiogram.handlers import BaseHandler
 from aiogram.types import Update
 
 pytestmark = pytest.mark.asyncio
