+++ tests/test_handler/test_chat_member.py
@@ -3,7 +3,7 @@ from typing import Any
 
 import pytest
 
-from aiogram.handler import ChatMemberHandler
+from aiogram.handlers import ChatMemberHandler
 from aiogram.types import Chat, ChatMemberMember, ChatMemberUpdated, User
 
 pytestmark = pytest.mark.asyncio
